+++ migrations/000042_mail_image_senders.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS mail.image_allowed_senders;